osm2railjson: move operational point code
Move operational point code to a separate file.
Add two new functions for op parts and trigrams for better readability.
