+++ editoast/osm_to_railjson/src/lib.rs
@@ -1,5 +1,6 @@
 mod generate_routes;
 mod generate_signals;
+mod operational_point;
 mod osm_to_railjson;
 mod utils;
